+++ packages/@expo/fingerprint/src/sourcer/Expo.ts
@@ -43,7 +43,7 @@ export async function getExpoConfigSourcesAsync(
     // icons
     expoConfig.icon,
     isAndroid ? expoConfig.android?.icon : undefined,
-    isIos ? expoConfig.ios?.icon : undefined,
+    ...(isIos ? collectIosIcons(expoConfig.ios?.icon) : []),
     isAndroid ? expoConfig.android?.adaptiveIcon?.foregroundImage : undefined,
     isAndroid ? expoConfig.android?.adaptiveIcon?.backgroundImage : undefined,
     expoConfig.notification?.icon,
@@ -219,6 +219,19 @@ function postUpdateExpoConfig(config: ExpoConfig, projectRoot: string): ExpoConf
   return config;
 }
 
+/**
+ * Collect iOS icon to flattened file paths.
+ */
+function collectIosIcons(icon: NonNullable<ExpoConfig['ios']>['icon']): string[] {
+  if (icon == null) {
+    return [];
+  }
+  if (typeof icon === 'string') {
+    return [icon];
+  }
+  return [icon.light, icon.dark, icon.tinted].filter((file): file is string => Boolean(file));
+}
+
 export async function getEasBuildSourcesAsync(projectRoot: string, options: NormalizedOptions) {
   const files = ['eas.json', '.easignore'];
   const results = (
@@ -259,6 +272,16 @@ export async function getExpoAutolinkingAndroidSourcesAsync(
         project.sourceDir = filePath; // use relative path for the dir
         debug(`Adding expo-modules-autolinking android dir - ${chalk.dim(filePath)}`);
         results.push({ type: 'dir', filePath, reasons });
+        // `aarProjects` is present in project starting from SDK 53+.
+        if (project.aarProjects) {
+          for (const aarProject of project.aarProjects) {
+            // use relative path for aarProject fields
+            aarProject.aarFilePath = toPosixPath(
+              path.relative(projectRoot, aarProject.aarFilePath)
+            );
+            aarProject.projectDir = toPosixPath(path.relative(projectRoot, aarProject.projectDir));
+          }
+        }
       }
       if (module.plugins) {
         for (const plugin of module.plugins) {
@@ -268,6 +291,7 @@ export async function getExpoAutolinkingAndroidSourcesAsync(
           results.push({ type: 'dir', filePath, reasons });
         }
       }
+      // Backward compatibility for SDK versions earlier than 53
       if (module.aarProjects) {
         for (const aarProject of module.aarProjects) {
           // use relative path for aarProject fields