+++ packages/expo-modules-core/ios/Core/Convertibles/Convertibles+Color.swift
@@ -80,6 +80,32 @@ extension UIColor: Convertible {
     throw Conversions.ConvertingException<UIColor>(value)
     // swiftlint:enable force_cast
   }
+
+  public static func convertResult(_ result: Any, appContext: AppContext) throws -> Any {
+    if let value = result as? UIColor {
+      return value.string
+    }
+    return result
+  }
+}
+
+extension UIColor {
+  public var string: String {
+    var r: CGFloat = 0
+    var g: CGFloat = 0
+    var b: CGFloat = 0
+    var a: CGFloat = 0
+
+#if os(macOS)
+    getRed(&r, green: &g, blue: &b, alpha: &a)
+    return String(format: "#%02x%02x%02x%02x", Int(r * 255), Int(g * 255), Int(b * 255), Int(a * 255) )
+#else
+    if getRed(&r, green: &g, blue: &b, alpha: &a) {
+      return String(format: "#%02x%02x%02x%02x", Int(r * 255), Int(g * 255), Int(b * 255), Int(a * 255) )
+    }
+#endif
+    return "#00000000"
+  }
 }
 
 extension CGColor: Convertible {
@@ -92,6 +118,12 @@ extension CGColor: Convertible {
       throw Conversions.ConvertingException<CGColor>(value)
     }
     // swiftlint:enable force_cast
+  }
+
+  public static func convertResult(_ result: Any, appContext: AppContext) throws -> Any {
+    let cgColor = result as! CGColor
+    let uiColor = UIColor(cgColor: cgColor)
+    return try UIColor.convertResult(uiColor, appContext: appContext)
   }
 }