+++ packages/expo-modules-core/jest-rsc.config.js
@@ -1,0 +1,1 @@
+module.exports = require('jest-expo/rsc/jest-preset');