+++ packages/expo-router/src/ui/TabTrigger.tsx
@@ -172,9 +172,10 @@ export function useTabTrigger(options: TabTriggerProps): UseTabTriggerResult {
           return router.navigate(config.href);
         } else {
           return navigation?.dispatch({
+            ...config.action,
             type: 'JUMP_TO',
             payload: {
-              name,
+              ...config.action.payload,
               ...options,
             },
           });