+++ packages/expo-updates/android/src/androidTest/java/expo/modules/updates/loader/RemoteLoaderTest.kt
@@ -72,7 +72,7 @@ class RemoteLoaderTest {
       directiveUpdateResponsePart = null
     )
 
-    coEvery { mockFileDownloader.downloadAsset(any(), any(), any(), any()) } answers {
+    coEvery { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) } answers {
       val asset = firstArg<AssetEntity>()
       FileDownloader.AssetDownloadResult(asset, true)
     }
@@ -85,7 +85,7 @@ class RemoteLoaderTest {
     }
 
     Assert.assertNotNull(result.updateEntity)
-    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -96,7 +96,7 @@ class RemoteLoaderTest {
 
   @Test
   fun testRemoteLoader_FailureToDownloadAssets() = runTest {
-    coEvery { mockFileDownloader.downloadAsset(any(), any(), any(), any()) } throws IOException("mock failed to download asset")
+    coEvery { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) } throws IOException("mock failed to download asset")
 
     try {
       loader.load { _ ->
@@ -106,7 +106,7 @@ class RemoteLoaderTest {
       Assert.assertEquals("mock failed to download asset", e.message)
     }
 
-    coVerify(atLeast = 1) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(atLeast = 1) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -133,7 +133,7 @@ class RemoteLoaderTest {
     Assert.assertNotNull(result.updateEntity)
 
     // only 1 asset (bundle) should be downloaded since the other asset already exists
-    coVerify(exactly = 1) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 1) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -162,7 +162,7 @@ class RemoteLoaderTest {
     Assert.assertNotNull(result.updateEntity)
 
     // both assets should be downloaded regardless of what the database says
-    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -196,7 +196,7 @@ class RemoteLoaderTest {
     }
 
     Assert.assertNotNull(result.updateEntity)
-    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -224,7 +224,7 @@ class RemoteLoaderTest {
     Assert.assertNotNull(result.updateEntity)
 
     // missing assets should still be downloaded
-    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -252,7 +252,7 @@ class RemoteLoaderTest {
     }
 
     Assert.assertNotNull(result.updateEntity)
-    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -278,7 +278,7 @@ class RemoteLoaderTest {
     }
 
     Assert.assertNotNull(result.updateEntity)
-    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(1, updates.size)
@@ -300,7 +300,7 @@ class RemoteLoaderTest {
 
     Assert.assertEquals(updateDirective, result.updateDirective)
     Assert.assertNull(result.updateEntity)
-    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 0) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     val updates = db.updateDao().loadAllUpdates()
     Assert.assertEquals(0, updates.size)
@@ -317,7 +317,7 @@ class RemoteLoaderTest {
       Loader.OnUpdateResponseLoadedResult(shouldDownloadManifestIfPresentInResponse = true)
     }
 
-    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any()) }
+    coVerify(exactly = 2) { mockFileDownloader.downloadAsset(any(), any(), any(), any(), any(), any()) }
 
     Assert.assertEquals(2, progressUpdates.size)
     Assert.assertEquals(0.5, progressUpdates[0], 0.001)